@@ -3,17 +3,13 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
 using SwissArmyKnife.Avalonia.Views;
 
-namespace SwissArmyKnife.Avalonia
-{
-    public class App : Application
-    {
-        public override void Initialize()
-        {
+namespace SwissArmyKnife.Avalonia {
+    public class App : Application {
+        public override void Initialize() {
             AvaloniaXamlLoader.Load(this);
         }
 
-        public override void OnFrameworkInitializationCompleted()
-        {
+        public override void OnFrameworkInitializationCompleted() {
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
                 desktop.MainWindow = new ProjectManagementWindow();
 
