@@ -1,7 +1,4 @@
-namespace SwissArmyKnife.Avalonia.Controls.G3D
-{
-    public class GLRenderer
-    {
-        
+namespace SwissArmyKnife.Avalonia.Controls.G3D {
+    public class GlRenderer {
     }
 }
