@@ -1,50 +1,25 @@
-﻿using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using Avalonia.Controls;
+﻿using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
-using BeaterLibrary.Formats.Pokémon;
-using SwissArmyKnife.Avalonia.ViewModels;
+using BeaterLibrary.Formats.Text;
+using BeaterLibrary.GameInfo;
+using SwissArmyKnife.Avalonia.Utils;
+using SwissArmyKnife.Avalonia.ViewModels.Editors;
 
-namespace SwissArmyKnife.Avalonia.Pages
-{
-    public class WildPokemonEditor : ReactiveUserControl<WildPokemonViewModel>
-    {
-        public WildPokemonEditor()
-        {
-            DataContext = new WildPokemonViewModel();
-            InitializeComponent();
+namespace SwissArmyKnife.Avalonia.Pages {
+    public class WildPokemonEditor : ReactiveUserControl<WildPokemonViewModel> {
+        public WildPokemonEditor() {
+            if (!Design.IsDesignMode)
+                DataContext = new WildPokemonViewModel();
+            initializeComponent();
         }
 
-        private void InitializeComponent()
-        {
+        private void initializeComponent() {
             AvaloniaXamlLoader.Load(this);
         }
 
-        public void HandleWildPokemonEncounter(bool Saving)
-        {
-            void BinaryToWildPokemonEncounter(string path)
-            {
-                DataContext = new WildPokemonViewModel(new WildEncounters(new BinaryReader(File.OpenRead(path))));
-            }
-
-            void WildPokemonEncounterToBinary(string path)
-            {
-                new WildEncounters
-                {
-                    GrassEntries = ViewModel.GrassEntries.ToList(),
-                    GrassDoubleEntries = ViewModel.GrassDoubleEntries.ToList(),
-                    GrassShakeEntries = ViewModel.GrassShakeEntries.ToList(),
-                    SurfEntries = ViewModel.SurfEntries.ToList(),
-                    SurfSpotEntries = ViewModel.SurfSpotEntries.ToList(),
-                    FishEntries = ViewModel.FishEntries.ToList(),
-                    FishSpotEntries = ViewModel.FishSpotEntries.ToList()
-                }.Serialize(path);
-            }
-
-            UIUtil.HandleFile(Saving, BinaryToWildPokemonEncounter, WildPokemonEncounterToBinary,
-                new List<FileDialogFilter>());
+        private void SelectingItemsControl_OnSelectionChanged(object? sender, SelectionChangedEventArgs e) {
+            throw new System.NotImplementedException();
         }
     }
 }
